Overpass-based augmented diff source + parser
The parser emits GeoJSON FeatureCollections w/ a top-level id
corresponding to the action (create/modify/delete). If a modify or
a delete, there will be 2 Features, each with an id of "old" or "new"
depending on the half of the change they correspond to.

Deletes retain the geometry of the previous version for findability.

Top-level changesets are changed to reflect the latest changeset that
affected the top-level element when it was not modified itself.

Note: only supports nodes + ways at present (osmtogeojson is an option
for parsing everything).